config-update to use many jobs with small number of pods
last SD bug was trigered with a small number of pods and a large number
of jobs.

this also allows us to use a single node for the fake-webserver.

This config is tested to utulize the fake-webserver node at 80% at 20
pods and the Prometheus nodes at around 60% avarage(we need some room
for the compaction processes).

Updated the scrape_interval: to 5s as this is closer to a real
production environment.

Signed-off-by: Krasi Georgiev <kgeorgie@redhat.com>