Upgrade load generator dependencies
This commit updates the `pyopenssl` dependency in the
`requirements.txt` file in the `load-generator` directory.
The reason for this update is to address an
`AttributeError` issue that was occurring due to the absence of the
`X509_V_FLAG_CB_ISSUER_CHECK` attribute in the `lib` module. The error
trace showed that this issue was specifically occurring in the
`urllib3.contrib.pyopenssl` module, thus necessitating the update of the
`pyopenssl` dependency to version `23.0.0`. Additionally, a few other
dependencies were also updated to their latest versions for
compatibility purposes.

Signed-off-by: Julien Pivotto <roidelapluie@o11y.eu>