Removed funcbench from this repository.
It was fun to develop it and use for a while. However, currently it was not used
for long (AFAIK), and definitely not maintained/updated for a while.

Other communities (e.g. Otel) started to use [dedicated OSS action](https://github.com/benchmark-action/github-action-benchmark) which looks promising.

I propose we remove the funcbench from our repo to not confuse users or our devs that such logic "should" work
on our repo.

Signed-off-by: bwplotka <bwplotka@gmail.com>

# Conflicts:
#	funcbench/bench.go
#	funcbench/env.go
#	funcbench/main.go
#	go.mod